Fix markup lists and inline image
Without this change it shows Lorem Lipsum placeholder text if a post
contains lists or inline images.

Right now it will display inline images as hyperlinks, but it should
probably be changed to download the image and display it inline.
That should be done in another task unless its simple to do.

I also tried to fallback to non markup mode for the label if pango fails
to parse the markup but pango fails to parse <a> tags, even though
setting <a> tags works if you set them directly with label = ...